+++ tests/test_pytest_plugin.py
@@ -676,3 +676,23 @@ class TestFreePortFactory:
         for family, addr in families:
             with socket.socket(family, socket.SOCK_DGRAM) as sock:
                 sock.bind((addr, free_udp_port))
+
+
+def test_func_as_parametrize_param_name(testdir: Pytester) -> None:
+    """
+    Test that "func" can be used as a parameter name in
+    `pytest.mark.parametrize` when using the pytest plugin.
+    """
+    testdir.makepyfile(
+        """
+        import pytest
+
+        @pytest.mark.parametrize("func", [1])
+        @pytest.mark.anyio
+        async def test_func_as_parametrize_param_name(func: int) -> None:
+            pass
+        """
+    )
+
+    result = testdir.runpytest(*pytest_args)
+    result.assert_outcomes(passed=len(get_available_backends()))